chore(route-utils): rename 'routeUtils.ts' to 'route-utils.ts'
