+++ packages/wp-source/index.ts
@@ -1,7 +1,7 @@
 import { Action, State } from "@frontity/types";
 import Source from "@frontity/source";
 import { EntityData } from "@frontity/source/src/data";
-import { Api, Resolver, routeUtils } from "./src/libraries";
+import { Api, Resolver } from "./src/libraries";
 
 export type Handler = (
   state: State<WpSource>["source"],
