+++ packages/wp-source/src/libraries/handlers/__tests__/date.tests.ts
@@ -6,7 +6,7 @@ import { mockResponse, expectEntities } from "./mocks/helpers";
 import posts2016 from "./mocks/posts2016.json";
 import posts2016p2 from "./mocks/posts2016p2.json";
 import posts20161025 from "./mocks/posts2016-10-25.json";
-import routeUtils from "../../routeUtils";
+import routeUtils from "../../route-utils";
 
 let state: State<WpSource>["source"];
 let libraries: WpSource["libraries"];
