+++ packages/wp-source/src/libraries/index.ts
@@ -1,4 +1,4 @@
 export { default as Api } from "./api";
 export { default as populate } from "./populate";
 export { default as Resolver } from "./resolver";
-export { default as routeUtils } from "./routeUtils";
+export { default as routeUtils } from "./route-utils";
